匹配接口

+++ contract/src/main/java/com/liepin/contract/entity/vo/list/GetContractsListVO.java
@@ -35,5 +35,5 @@ public class GetContractsListVO {
     @ApiModelProperty(notes = "备注")
     private String remark;
     @ApiModelProperty(notes = "已匹配人数")
-    private Integer matchNum;
+    private Long matchNum;
 }
